xgboost_v3/requirements.txt
# Core ML libraries
xgboost>=2.0.0
scikit-learn>=1.3.0
optuna>=3.4.0
imbalanced-learn>=0.11.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0  # для parquet

# Database
psycopg2-binary>=2.9.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
joblib>=1.3.0

# GPU support (optional)
# Uncomment if using GPU
# GPUtil>=1.4.0

# Additional for advanced features
scipy>=1.10.0
statsmodels>=0.14.0